limit countBy methods to return long instead of int

+++ tck/src/main/java/ee/jakarta/tck/data/standalone/nosql/example/Catalog.java
@@ -28,8 +28,8 @@ public interface Catalog extends DataRepository<Product, Long> {
     void save(Product product);
     void deleteById(Long id);
     
-    int countByPriceGreaterThanEqual(Double price);
-    int countBySurgePriceGreaterThanEqual(Double price);
+    long countByPriceGreaterThanEqual(Double price);
+    long countBySurgePriceGreaterThanEqual(Double price);
     
     List<Product> findByNameLike(String name); 
 }
